+++ src/lib.rs
@@ -149,6 +149,7 @@ pub(crate) use archparam_defaults as archparam;
 
 mod gemm;
 mod kernel;
+mod packing;
 mod ptr;
 mod threading;
